+++ changelog.md
@@ -7,6 +7,10 @@
 -   Private `CircularCountdown` component
 -   `Notification` component
 
+### Changed
+
+-   Use kebab case for close reasons (Overlay, Dialog)
+
 ## [8.0.0-alpha.5] - 2019-04-22
 
 ### Fixed